modules/generic_stage_target.py: handle USE_EXPAND flags
Currently catalyst does not handle USE_EXPAND flags.  However, the
recent move of mmx, sse and friends from the global USE flags to the
new CPU_FLAGS_X86 USE_EXPAND flag requires this.

This commit adds code to handle a new syntax for self.setting[x] values
which can now contain dictionaries to objects.  We use these to build
environment variables for USE_EXPAND flags.  Eg. in arch/amd64.py
we set

 self.settings["HOSTUSEEXPAND"]={"CPU_FLAGS_X86":["mmx","sse","sse2"]}

which is later exported in the environtment as:

    clst_HOSTUSEEXPAND="CPU_FLAGS_X86"

and

    clst_CPU_FLAGS_X86="mmx sse sse2"

In general one could have:

  self.settings["A"]={"B":[1,2,3],"C":[4,5,6]}

which would export

 clst_A="B C"

and

 clst_B="1 2 3"
 clst_C="4 5 6"

Signed-off-by: Anthony G. Basile <blueness@gentoo.org>