targets/stage1/stage1-chroot.sh: don't quote ${clst_HOSTUSEEXPAND}
Remove the quotes on ${clst_HOSTUSEEXPAND} so the for loops skip
any iterations when clst_HOSTUSEEXPAND is unset.  Otherwise a
spurious line is introduced in the make.conf file which causes
a syntax error.

To illustrate the problem, run the follow two lines of bash:

  unset A ; for i in $A; do echo x${i}x ; done    #line 1
  unset A ; for i in "$A"; do echo x${i}x ; done  #line 2

Line 1 leads to no output, while line 2 leads to "xx".

We also fix a typo in the sed line.

Signed-off-by: Anthony G. Basile <blueness@gentoo.org>