use profile for CPU_FLAGS_X86 where appropriate
a number of the CPU_FLAGS_X86 settings appeared to be trying to use something
very much like the default from the profile but missing mmxext.  Due to this,
a number of packages like ffmpeg would be broken, where using the profile is
better.

Modified to remove (comment out) the ones that appeared to be attempting to
use the default (and failing) while leaving in the settings which are intentionally
disabling certain unavailable features per sub-arch.