pylint: add a helper for linting code
A bunch of warnings are left disabled because there's too many violations
in the current code base.  Hopefully we can get to a point where everyone
uses this though and we can start enabling more and more.

The code base is "clean" right now in the sense that running pylint will
not produce any output.  But we should start removing violations one by
one and re-enable the warning class as we go.