support: punt pointless list_to_string helper
This function is a lot of code that doesn't do a whole lot.
It looks like a blind copy & paste & tweak from list_bashify.
If the input is a string, it returns a copy of it (i.e. it's
pretty much a nop).  If it's a list, it runs ' '.join() on it.
That's all.

But let's not focus on the craziness of implementation.  The
few consumers of this function are already passing in strings
in which case their use is pointless.  Once we delete those
callers, we can delete the implementation.