Catalyst: Fix regression from commit: 1d5d16c965 Remove hard coded portdir paths
Convert all clst_portdir use to the new clst_repo_basedir and clst_repo_name variables.
This de-couples the target setting from the source portdir setting for the source repo.
Without this if the source repo was not the same as theses two combined, it would fail to find
it's target scripts and information.