base/stagebase.py: Cleanup and fix the autoresume validation in the unpack()
The previous logic was convoluted and had many errors, most notably, not checking
the chroot is a directory, but instead was checking the seed source.
I broke up and simplified the logic into two sections, one checks the chroot for a valid resume
point, the other checks for seed tarball changes which would invalidate the existing chroot.

Also the non-autoresume code block code was mostly not needed and was also simplified.