create-iso.sh: use --apparent-size when calculating ESP size
When catalyst uses du to check the size of files to be stored in the
vfat loopback image, FS compression (on the catalyst temp space) may
cause it to report a smaller value than will be needed on vfat.  This
results in "No space left on device" when populating the EFI image.

Adding --apparent-size will tell the true size (rather than on-FS size)
which should be closer to what's needed when stored on vfat.