targets: Reduce locales to C.UTF8 in stage builds
By default, glibc generates around 500 locales with more added each
year.

With USE=-compile-locales, glibc generates the locale archive in
pkg_postinst(). Since files generated in pkg_postinst() are not recorded
in the vdb, this has the advantage of allowing users to freely change
the set of enabled locales (by editing /etc/locale.gen and running
locale-gen).

Since it is so easy for the user to generate any locales they want with
locale-gen (and they probably would have anyway to rid themselves of the
499 locales they don't want!), just disable all locales except for
C.UTF8 and save stage builders a lot of time.

The patch works by
	(1) Writing /etc/locale.gen with "C.UTF8 UTF-8"
	(2) Setting CONFIG_PROTECT so glibc doesn't overwrite
	    /etc/locale.gen
	(3) Running etc-update to reset /etc/locale.gen

In order to do this I modified scripts/bootstrap.sh in commit
0aa49828ae25 (scripts/bootstrap.sh: Allow CONFIG_PROTECT).

Reducing the set of locales cuts the user time (as reported by time(1))
of the stage2 and stage3 builds as well as the file size of the
resulting xz'd tarballs:

            stage 2           stage 3
         size     time     size      time
before    89M   22m42s     206M     45m5s
 after    77M    4m29s     195M     26m8s

An alternative solution would be to set USE=compiled-locales for glibc,
but that has the downside of being non-default and likely causing users
to unnecessarily rebuild glibc. (We'll do this for the ISOs where we
want all the locales)

Note that this patch does not change the contents of /etc/locale.gen in
the stage3 tarball.

Closes: https://bugs.gentoo.org/686862
Signed-off-by: Matt Turner <mattst88@gentoo.org>
(cherry picked from commit 5fb710093c0d4643b981db7763f7f14d04e02d79)