catalyst: Enable and fix CPU_FLAGS_X86
In commit be6d7f0b7f0a (use profile for CPU_FLAGS_X86 where appropriate)
we disabled setting CPU_FLAGS_X86 to avoid overriding the profile's
defaults (and in some cases, overriding incorrectly).

While that is undoubtedly the correct thing to do for the generic
classes (e.g., arch_amd64) it does not make sense for the classes
specific to a particular microarchitecture, unless we expect the
defaults of CPU_FLAGS_X86 to change in the future, which seems highly
unlikely.

It also is the wrong thing to do for x86, which does not have a default
CPU_FLAGS_X86.

Fixes: be6d7f0b7f0a (use profile for CPU_FLAGS_X86 where appropriate)
Signed-off-by: Matt Turner <mattst88@gentoo.org>