targets: Use GRUB for BIOS boot
grub-mkrescue produces an ISO that is bootable from EFI as well as from
the BIOS, so isolinux isn't necessary to accomplish that. The benefits
are that (1) we reduce the number of dependencies required to make the
ISO, (2) the same GRUB set up works for EFI and BIOS systems, and (3)
grub-mkrescue replaces a whole lot of code that we no longer need to
maintain.

This patch has the side effect of removing the F[2-7] help texts and the
memtest86+ option, since those were done via isolinux. I feel okay
removing those since isolinux was only used on BIOS systems and not on
EFI, so none of those things were available to the vast majority of our
users.

It also has the side-effect of removing Fedora's signed UEFI shim
(sys-boot/shim) which could be used to add a hash for the live CD's grub
binary so that you could install under Secure Boot. That doesn't offer
any protection, and the author of the code indicated to me that the
value in offering this feature is that it allows installation on systems
with broken firmware under which important hardware (like the NVMe
drive) isn't exposed unless using Secure Boot. After quite a bit of
searching the internet, I can't find any mentions of this being a
problem.

Signed-off-by: Matt Turner <mattst88@gentoo.org>