catalyst: Remove 'contents' config option
This option seems completely absurd and on top of that has been
completely broken, since contents() has been passed 'j', which is the
mode, as the positional argument 'destination'. This only worked because
destination is not used in the contents generation and the actual mode
parameter has a default argument of 'auto'.

So the config option has not worked in a long time and no one has
noticed.

Signed-off-by: Matt Turner <mattst88@gentoo.org>