catalyst: Remove remnants of options="preserve_libs"
This option has been broken since commit e3114e65ffda (Some options
cleanup, unifying their use, reducing redundancy.) in 2013 which was
first in catalyst 3.0.0.

I think the feature was a kludge, and that if or when we find the need
for something similar that we instead just run emerge @preserved-rebuild
instead.

Signed-off-by: Matt Turner <mattst88@gentoo.org>