embedded: use destpath to clean mergeroot
When using remove() with a capture subpath (e.g. the embedded target), specifying
files in rm removes the files from the seed stage, not the captured stage.

Use destpath as a base for removals. When not using a subpath, destpath
equals chroot_path.

Also unhardcode log prefix to match the target.

Signed-off-by: Matt Turner <mattst88@gentoo.org>