targets: Remove ancient locale-deleting code
This block of code was added in commit ef648688ce8b ("Forward port the
changes from catalyst 1.1.9 to 1.1.10.10 to catalyst2.") in 2005 without
any further explanation. Since locale-gen was first committed to our
glibc package in 2006, we think that locales were handled differently
before (and not packed into /usr/${libdir}/locale/locale-archive).

Not only is this code unneeded, it can actually cause the locale-archive
to be deleted if the livecd is using a multilib profile! E.g. on a
multilib amd64 profile, where /usr/lib/locale is a symlink to
/usr/lib64/locale, it deletes /usr/lib64/locale/l* which matches
/usr/lib64/locale/locale-archive.

Because the minimal livecd uses a non-multilib profile, this explains
why it still contained the expected set of locales, while the multilib
admincd did not.

Thanks go to Ben Kohler, Thomas Deutschmann, and  Andreas Hüttel for
figuring out this mystery!

Bug: https://bugs.gentoo.org/710762
Signed-off-by: Matt Turner <mattst88@gentoo.org>