targets: Use pkgcache for kernel sources
The rationale for not using it before made sense, given that multiple
binpkgs could not exist for the same version of a package. But we now
use FEATURES=binpkg-multi-instance so this is no longer the case.

The dumbest part of this is that the code wasn't even disabling the
package cache. It was instead just overriding PKGDIR to dump the binary
package into a location (named "ebuilds" no less!) that emerge would
never otherwise use.

Signed-off-by: Matt Turner <mattst88@gentoo.org>