targets: Default to sys-kernel/gentoo-sources directly
Using virtual/linux-sources was a nice idea, but leads to multiple
failures in practice. For example, we use 'portageq contents' later, and
the virtual provides no files. We could handle the indirection with
'portageq expand_virtual' in some cases but that requires the virtual to
be installed on the system.

Fixes: 388def77 ("kmerge.sh blows away the /usr/src/linux symlink ...")
Signed-off-by: Matt Turner <mattst88@gentoo.org>