targets: Remove wrong make.conf munging
With pkgcache fixed for kernel sources in a prior commit, I noticed that
if a livecd-stage2 built three kernels, the first would emerge normally
and build a binpkg, the second would emerge and fail to use the binpkg
produced earlier, and the third would use a previously produced binpkg.
I discovered the reason is that the second and third emerges are run
after USE="... build" has been added to make.conf. We intentionally
removed this in commit c0fcbf02 ("kmerge.sh: Don't set USE=build for
emerging kernels.") years ago, so I was surprised to find this code.

The commit mentioned below inexplicably (and wrongly) readded USE=build
to make.conf, failed to update the sed command to remove it, and readded
the addition in entirely the wrong spot so it didn't even affect the
kernel merge.

genkernel doesn't source the system's make.conf, so this code would be
entirely useless if it wasn't harmful.

Fixes: 0343a18e ("Fix merging kernel without kerncache enabled.")
Signed-off-by: Matt Turner <mattst88@gentoo.org>