arch/ppc.toml: don't force -mabi=elfv2 on power subarches.
-mabi=elfv2 is default on ppc64le in gcc/glibc toolchain for all
powerpc64le-*-linux-gnu CHOSTS, so we don't need to pass it.

forcing -mabi=elfv2 on ppc64(be) is quite risky, as it's not standard
configuration and not officialy supported by glibc at all.
according to glibc upstream it should be a separate ABI port.
Also many apps/tools assume elfv1 on big-endian, and while fixes are
being upstreamed, we've observed breakage in: openssl[asm], nodejs,
ffmpeg and  others.

If we want to ship elfv2 ppc64be stage,
we should define it as separate subarch and strictly put it in experimental.

to my knowledge those subarch definitions are not used by releng yet, so
we will not break anything.

Signed-off-by: Georgy Yakovlev <gyakovlev@gentoo.org>
Signed-off-by: Matt Turner <mattst88@gentoo.org>