arch: add new x86.i686-ssemath subprofile
As mentioned on the mailing list, x86 arch testing needs to be performed
using SSE registers for floating-point math instead of the 387 FPU.
This is a subprofile which provides a generic i686 target with SSE2 and
`-mfpmath=sse`.  The use is not just limited to arch testing however as
this is a useful platform for getting maximum performance on late-gen
chips.  Per gcc:

> The resulting code should be considerably faster in the majority of
cases and avoid the numerical instability problems of 387 code, but may
break some existing code that expects temporaries to be 80 bits.

Of course modern code has quite the opposite problem, where it does not
expect floating-point variables to be 80 bits.

Signed-off-by: Matoro Mahri <matoro_gentoo@matoro.tk>
Closes: https://github.com/gentoo/catalyst/pull/12
Signed-off-by: Ben Kohler <bkohler@gentoo.org>