Add x86-64-v3 baseline support
Adds x86-64-v3 baseline support with and without LTO to Catalyst.
While not useful to Gentoo right now it will likely become something we add in the future seeings as we already have binhost support for it, so will save someone having to do the work again when that time comes.

Signed-off-by: Ian Jordan <immoloism@gmail.com>