catalyst/*: add new option iso_extra_partition
When specified in the spec, this will instruct create-iso.sh to
create an extra XFS partition of the specified size.

The intended use is having a partition available for persistent
storage when the created hybrid ISO image is written to e.g. a USB.

When the hybrid ISO is burned to an optical medium the extra
partition is ignored, along with all other MBR data.

Signed-off-by: Nowa Ammerlaan <nowa@gentoo.org>