update_seed_command: revert to a boring but safe default
The changed-subslot logic sounds nice, but unfortunately does not work all
the time for reasons not yet known. Surprise fun and breakage ensues.

These days we're more concerned that catalyst can run unsupervised for long
times and less that it finishes in record cpu time, so set the default
to an update command which we also recommend to all our users and know to
be robust.

Signed-off-by: Andreas K. Hüttel <dilfridge@gentoo.org>